[generator] Rename internal files
Rename files to match the method names. Does not change code or public facing files.
