+++ graphql-kotlin-schema-generator/src/test/kotlin/com/expediagroup/graphql/generator/state/ClassScannerTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Expedia, Inc
+ * Copyright 2020 Expedia, Inc
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package com.expediagroup.graphql.generator
+package com.expediagroup.graphql.generator.state
 
-import com.expediagroup.graphql.generator.state.ClassScanner
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
