+++ graphql-kotlin-schema-generator/src/test/kotlin/com/expediagroup/graphql/generator/types/GenerateScalarTest.kt
@@ -28,7 +28,7 @@ import kotlin.reflect.full.createType
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 
-internal class ScalarBuilderTest : TypeTestHelper() {
+internal class GenerateScalarTest : TypeTestHelper() {
 
     internal class Ids {
         internal val stringID: String = "abc"
